Fix DetachError issue
After recent changes to retry less times on DetachForFailure, this error was being hit from time to time and would overwrite the error object on the release which was responsible for sending the deploy into the failure state in the first place. As we don't care to wait for detach to finish here, we increase the wait after this to 60s. Other errors in DetachForFailure would still behave the same way, which seems like what we'd want here as they are not expected like this one.