Use latest version of `step` with DynamoDB root locking
We are moving from S3 to DynamoDB for locking as the former is eventually consistent.
The upstream changes in `step` have already been made: https://github.com/coinbase/step/pull/55

This commit updates `odin` to use the latest version of `step`
with DynamoDB root locking. By default, we are using the name
of the deployed lambda postfixed with `-locks` as the DynamoDB
lock table name.

TODO:

- Some testing
- Update go mods